fix: widen sub-account index to u32 to prevent JS truncation past validation
wasm-bindgen narrowed a JS number to u8 *before* the > 31 check, so e.g. 257
wrapped to 1 and returned an address instead of throwing. Take index as u32
(at both the wasm boundary and the core fn), validate the full value, and carry
it as u32 in SDKError::InvalidSubAccountIndex. Tests cover 257/288 wrap cases.

+++ rust/src/errors.rs
@@ -138,9 +138,11 @@ pub enum SDKError {
     InvalidMultisig(String),
 
     /// Sub-account index out of range. The runtime tracks sub-account
-    /// existence in a `u32` bitmask, so valid indices are `0..=31`.
+    /// existence in a `u32` bitmask, so valid indices are `0..=31`. Carries the
+    /// caller-supplied value as `u32` so an out-of-`u8` value (e.g. from a JS
+    /// number) is reported faithfully rather than truncated.
     #[error("Sub-account index {0} out of range (0..=31)")]
-    InvalidSubAccountIndex(u8),
+    InvalidSubAccountIndex(u32),
 
     #[error(transparent)]
     WebsocketError(#[from] Box<WSErrors>),
